Move Storage into ModelManager
This is because we want to use the filesystem state as the canonical
representation sometimes, i.e. for article storage, which we don't want
to store in memory as the articles can be quite large.

We could simply expose the Storage to the Model as well, but doing so
results in Storage being referenced both from the outside at Logic,
and also from the inside whenever articles have to be added.

Instead we move the entire Storage behind Model for a cleaner
architecture -- now Storage should be an "implementation detail"
that the Model uses to persist itself.