Implement delete command for archives context
DeleteArchiveEntryCommand and DeleteArchiveEntryCommandParser classes created to implement this.
Using the same delete command and parser as the list-context was decided against.
The parser needs to return a specific Command object, so if we were to use the same parser, we need to use the same DeleteCommand class for deleting in both contexts.
This means that we have to rely on ModelManager#context to decide which entrybook to delete the entry from.
Could work, but feels very iffy.